requirement.txt
android==1.0
asgiref==3.2.7
astroid==2.2.5
audiostream==0.2
Django==3.0.6
djangorestframework==3.11.0
isort==4.3.15
jedi==0.13.2
Kivy==1.11.0.dev0
lazy-object-proxy==1.3.1
mccabe==0.6.1
parso==0.3.4
pygame-sdl2==2.1.0
pyjnius==1.1.5.dev0
pylint==2.3.1
PySDL2==0.9.4
pytz==2020.1
six==1.12.0
sqlparse==0.3.1
typed-ast==1.3.1
wrapt==1.11.1
